requirements.txt
docx2txt
langchain
langchain-core
langchain-community
langchain-groq
langgraph
chromadb
llama-index
llama-parse
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
python-dotenv
groq
